+++ src/main/java/com/calculator/io/Output.java
@@ -1,5 +1,5 @@
 package com.calculator.io;
 
 public interface Output {
-    void output(int result);
+    void output(double result);
 }
